pyproject.toml
[project]
name = "vinas-external-api-app"
version = "0.1.0"
description = ""
authors = [
    {name = "peyro", email = "peyrovskaaa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "redis (>=6.2.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-problem (>=0.11.4,<0.12.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "grpcio (>=1.76.0,<2.0.0)",
    "protobuf (>=6.33.0,<7.0.0)",
    "fastapi-pagination (>=0.15.0,<0.16.0)",
    "authtools @ git+https://github.com/vinasLT/shared_auth.git@v0.1.8",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "aio-pika (>=9.5.7,<10.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.2.0,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
